+++ packages/core/src/mechanic.js
@@ -14,8 +14,11 @@ import {
   htmlToCanvas,
   extractSvgSize,
   dataUrlToCanvas,
-  getTimeStamp
+  getTimeStamp,
+  mergeWithDefaultSettings
 } from "./mechanic-utils.js";
+
+import { mechanicDrawLoop } from "./mechanic-drawloop.js";
 
 import { MechanicError } from "./mechanic-error.js";
 
@@ -87,10 +90,15 @@ export class Mechanic {
       }
     }
 
-    this.settings = settings;
+    this.settings = mergeWithDefaultSettings(settings);
     this.values = values;
     this.functionState = lastRun?.functionState ?? {};
     this.exportType = exportType;
+
+    this.drawLoop = mechanicDrawLoop.prepare(this.settings.frameRate);
+
+    this.engineFrameCallback = () => {};
+    this.engineDoneCallback = () => {};
   }
 
   /**
@@ -142,16 +150,90 @@ export class Mechanic {
   }
 
   /**
+   ** Convenience function to collect all callbacks that an engine can pass to a
+   * design function. An engine can pass overwrites to inject custom behavior
+   * into the callbacks. This is currently used by engine-p5 to disable using
+   * mechanic.drawLoop within engine-p5.
+   *
+   * @param {object} overwrites - An object with overwrites for the callbacks
+
    * Returns an object with common functions to be used in the design function
    * @param {function} frame - The frame function
    * @param {function} done - The done function
    */
-  callbacks(frame, done) {
-    return {
-      frame,
-      done
-      //requestAnimationFrame:
+  callbacksForDesignFunction(overwrites = {}) {
+    const frame = (...args) => this.engineFrameCallback(...args);
+    const done = (...args) => {
+      // Always make sure to stop the drawloop when we're done rendering
+      this.drawLoop.stop();
+      this.engineDoneCallback(...args);
     };
+
+    return Object.assign(
+      {},
+      {
+        frame,
+        done,
+        // This is where the timeline feature could intercept to render a
+        // single frame. A single frame number can be passed as the second
+        // arguments to drawLoop's start method.
+        drawLoop: cb => this.drawLoop.start(cb),
+
+        // Keeping setState and state in their own namespace to avoid namespace
+        // collisions with React
+        mechanic: {
+          setState: this.setState.bind(this),
+          state: this.functionState,
+
+          /**
+           * Add a reference to the frame callback to the mechanic
+           * object for backwards compatibility.
+           *
+           * @deprecated
+           */
+          frame: (...args) => {
+            console.warn(
+              "mechanic.frame() has been deprecated and will be removed in a future version of Mechanic. The done() and frame() callbacks are now passed to your design function directly."
+            );
+            frame(...args);
+          },
+
+          /**
+           * Add a reference to the done callback to the mechanic
+           * object for backwards compatibility.
+           *
+           * @deprecated
+           */
+          done: (...args) => {
+            console.warn(
+              "mechanic.done() has been deprecated and will be removed in a future version of Mechanic. The done() and frame() callbacks are now passed to your design function directly."
+            );
+            done(...args);
+          }
+        }
+      },
+      overwrites
+    );
+  }
+
+  /**
+   * Helper function that sets the engineFrameCallback to an instance variable.
+   * Done as a function so validation and other setter logic can be added later.
+   *
+   * @param {function} frameCallback - The frame callback function
+   */
+  registerFrameCallback(frameCallback) {
+    this.engineFrameCallback = frameCallback;
+  }
+
+  /**
+   * Helper function that sets the engineFinalizeCallback to an instance variable.
+   * Done as a function so validation and other setter logic can be added later.
+   *
+   * @param {function} doneCallback - The done callback function
+   */
+  registerDoneCallback(doneCallback) {
+    this.engineDoneCallback = doneCallback;
   }
 
   /**
@@ -183,7 +265,7 @@ export class Mechanic {
       this.serializer = new XMLSerializer();
       this.videoWriter = new WebMWriter({
         quality: 0.95,
-        frameRate: 60
+        frameRate: this.settings.frameRate
       });
     }